Shift definition GrackleCtxPack to its own file
