+++ tests/grtestutils/CMakeLists.txt
@@ -80,6 +80,7 @@ add_library(grtestutils
   # files outside of the googletest subdirectory
   # -> these shouldn't include headers from the googletest subdirectory
   cmd.hpp cmd.cpp
+  GrackleCtxPack.hpp
   iterator_adaptor.hpp
   os.hpp os.cpp
   preset.hpp preset.cpp
