+++ tests/grtestutils/googletest/fixtures.hpp
@@ -6,7 +6,7 @@
 //===----------------------------------------------------------------------===//
 ///
 /// @file
-/// Define machinery for creating GoogleTest Fixutures to help test Grackle's
+/// Define machinery for creating GoogleTest Fixtures to help test Grackle's
 /// C API.
 ///
 //===----------------------------------------------------------------------===//
@@ -18,6 +18,7 @@
 // because we include gtest.h here, we should NOT include this file in any
 // grtest source files (in other words, this should be a header-only file)
 
+#include "../GrackleCtxPack.hpp"
 #include "../preset.hpp"
 
 namespace grtest {
