+++ tests/grtestutils/iterator_adaptor.hpp
@@ -16,7 +16,7 @@
 #include <string>
 
 #include "grackle.h"
-#include "preset.hpp"
+#include "GrackleCtxPack.hpp"
 
 namespace grtest {
